+++ src/App/Documentation/core/Color/index.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
 
-import Color, { ColorPalette, Usage, ColorText } from "./index";
+import Color, { ColorPalette, ColorText } from "./index";
 
 describe("Core: Color", () => {
     it("is defined", () => {
@@ -19,14 +19,6 @@ describe("Core: Color", () => {
     it("ColorPalette renders", () => {
         const wrapper = shallow(
             <ColorPalette />
-        );
-
-        expect(wrapper).toMatchSnapshot();
-    });
-
-    it("Usage renders", () => {
-        const wrapper = shallow(
-            <Usage />
         );
 
         expect(wrapper).toMatchSnapshot();